WIP: tests: import pkgconf tests
There are still some tests missing in the toml (and some tests that need
to be clarified with upstream because they exist but don't run and don't
pass). There are also just a lot of legitimate failures here we need to
sort out.

Also, because these are mostly broken anyway I haven't added this suite
to the CMake build, if someone wants to do that I can squash that change
into this.